requirements.txt
Faker==16.6.0
pandas==1.5.2
psycopg2-binary==2.9.5